Move share button into its own component
